requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
langchain
langchain-google-genai
langchain-groq
pydantic
